Update Scala Native to 0.5 and drop Scala 2.11
