+++ .github/workflows/ci.yml
@@ -24,7 +24,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest]
-        scala: [2.11.12, 2.12.17, 2.13.11]
+        scala: [2.12.19, 2.13.14]
         java: [temurin@8]
         platform: [jvm, js, native]
     runs-on: ${{ matrix.os }}
@@ -77,7 +77,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        scala: [2.13.11]
+        scala: [2.13.14]
         java: [temurin@8]
     runs-on: ${{ matrix.os }}
     steps:
