+++ README.md
@@ -120,20 +120,20 @@ resolvers ++= Seq(
 [ci]: https://travis-ci.org/milessabin/shapeless
 
 
-Builds are available for Scala 2.11.x, 2.12.x and 2.13.x. The main line of development for
-shapeless 2.3.3 is Scala 2.13.2.
+Builds are available for Scala 2.12 and 2.13. 
+The main line of development for shapeless 2.3.x is Scala 2.13.
 
 ```scala
-scalaVersion := "2.13.5"
+scalaVersion := "2.13.14"
 
 libraryDependencies ++= Seq(
-  "com.chuusai" %% "shapeless" % "2.3.3"
+  "com.chuusai" %% "shapeless" % "2.3.11"
 )
 ```
 
 For using snapshots of Shapeless you should add,
 ```scala
-scalaVersion := "2.13.5"
+scalaVersion := "2.13.14"
 
 libraryDependencies ++= Seq(
   "com.chuusai" %% "shapeless" % "2.4.0-SNAPSHOT"
@@ -163,8 +163,7 @@ releases][olderusage] on the shapeless wiki.
 
 ## Building shapeless
 
-shapeless is built with SBT 1.3.10 or later, and its main branch is built with Scala 2.13.2 by default but also
-cross-builds for 2.11.12 and 2.12.12.
+shapeless is built with SBT 1.x and its main branch is built with Scala 2.13 by default but also cross-builds for 2.12.
 
 [namehashing]: https://github.com/sbt/sbt/issues/1640
